Committing 99% Complete Text Editor Clone
The Find and Replace functionality has been completed and bug-fixed. The biggest problem was trying to get the Ctrl+Shift functionality to work, which never did, so an alternative was found. Now instead of Ctrl+Shift+F finding backwards, it is Ctrl+D, IE the key just prior to F. This is also the case with Replace, using Ctrl+E instead of Ctrl+Shift+R to replace and seek backwards for next one. Almost everything else is working as expected, including zoom functionality, translations, the swap button, etc.

There are only a very few things needed still, which I won't do before publishing:
* I need to actually detect line endings and do something with them. Currently it literally only shows the static text "UNIX (LF)" for line ending type, and there are no abilities to change that. This needs to work properly.
* The same actually goes for the text format detection and changing as well. Right now the widget and everything automatically assumes UTF-8, and if you aren't using UTF-8, you are chuck out of luck.
* While Translations technically "work," IE if they are changed and built it will show translated text, no actual text translations have been added.

These should get fixed later on as time permits, but this text editor is already pretty feature complete and competitive with Notepad, so it is a good time to publish it.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>